Move json.Number handler functions to api
